APREPRO: Added aprepro2 to replace aprepro
aprepro2 uses the aprepro_lib library instead of the internal
parsing and lexing of aprepro.  With this, then aprepro2 uses
the same lexing and parsing as everything else using the
aprepro_lib library (sierra, cubit, dakota, ...) which should
make compatibility easier to maintain and reduces code to maintain.

Still not ready to replace aprepro with aprepro2, but currently
testing what is different between the two.

Added 'readline' capability to aprepro_lib which provides
line recall and edittting for interactive use.