EXODUS_FOR: Add libexoIIv2for32 library
The default libexodus_for which is aliased to libexoIIv2for
is built assuming that all ints and reals are promoted to 8-byte
values by the compiler (for example -fdefault-real-8 -fdefault-int-8
for gfortran).  This library works well for linking to the seacas
fortran executables which are also compiled with these flags.

However, there are some fortran applications where the application
wants to use 4-byte integers and 8-byte or 4-byte reals and have
explicit control over this.  For example, the old pronto3d and jas3d
codes did this.

This library interface to exodus is provided by the libexoIIv2for32
library. The name of this may be confusing, but is retained for
backward compatibility with the old seacas version.

It is currently only built in SEACAS, not Trilinos.