SIMPI: Include mpi headers in include/siMPI
The pseudo-mpi implementation siMPI installs mpi.h and related
headers. Previously, these were installed in the default include
directory which could cause issues due to a subsequent parallel
compilation picking up the incorrect siMPI mpi.h instead of the true
mpi.h.

This commit moves the install directory for these siMPI mpi-related
headers to the siMPI subdirectory of the specified include directory.

Addresses Issue #14.  Needs to be propogated to the Trilinos siMPI,
but doing it here for testing purposes to make sure it will work as
intended.