EXODUS: Detect corruption in df count vs ss node list len
Fix due to below issue:

It looks like the distribution factor (df) count for sideset 6 is
incorrect.  There should be a one-to-one correspondence between
sideset df and the sideset nodes.  In this file, the df count for
sideset 6 is 15,104, but the number of nodes in the sideset
(basically, sum of number of nodes/face for each face in the sideset)
is 17,984.

There is currently an "optimization" or "kluge" in the
ex_get_side_set_node_list_len function where if the exodus file
contains a count for the distribution factors on a sideset, the
function will return that count instead of doing all the work to sum
up the nodes/face for each face.  This of course assumes that the df
count is correct.  If I remove that "optimization", then the function
returns 17,984 which is the same value that is calculated by the
ex_get_side_set_node_list function.

So, the file has incorrect information which is causing the memory
overwrite.  There are a few options for the exodus library.

* Remove the "optimization" and always calculate the node list length
the "correct" way.  If this count does not match the df count, then
output a warning message indicating that the file is likely corrupt.
This has the disadvantage of doing more work than should be needed for
"non-corrupt" files.

* Alternatively, the ex_get_side_set_node_list function could check
whether the calcuated length matches the df count and if not, then it
could output the likely corrupt message.  The question here is whether
it should then continue to fill the passed in array with the nodes and
likely segfault, or whether it should return with an error.

* Another option is to do the check in both locations. (THIS CHOSEN)

I'm not too concerned about the increase in computation time caused by
removing the optimization since that function is not used too much by
most codes.  If it becomes a bottleneck, there are some other safe
optimizations that can speed it up somewhat.