EXODUS, IOSS: Workaround netcdf bug.
There is a bug in netcdf-4.5.1-devel and earlier in which the code
will hang if parallel output is done for a element block attributes
if there are more than 1 attributed and mpiio (netcdf-4 hdf5-based)
collective output is used.

This fix changes the attribute variable type to INDEPENDENT if the
attribute count > 1 and parallel mpiio output is selected.

The bug is described in netcdf issues #447 and #448.