EXODUS: Ensure internal datastructures consistent
If the nc_close() or nc_sync() failed, the ex_close() would
return early and not update internal datastructures. If the
code then called ex_open or ex_create(), it was possible that
the datastrutures would be corrupt.

The change always cleans up the internal datstructures in ex_close()
even if nc_sync() or nc_close() fail.  The errors are still reported.