EPU: Add automatic subcycling option.
If -auto is specified and processor count exceeds max number of
open files, then the code will automatically activate the
-subcyle and -join_subcycle options.  The subcyle count will
be the cube root of the procesor count unless that exceeds
max_open_files in which case sqrt will be used. If that still
exceeds, will not activate subcyling.

There is a debug option -max_open_files which can be used to
set the max open file count to a smaller value -- useful for debugging