EPU: Preliminary "parallel" support
At this time, if epu is run on more than one processor,
it will split the model to be joined into "epu_proc_count"
subpieces and join each of the subpieces on a single processor.

If -join_subcycles is also specified, then when all processors are
finished, a single processor will then run and join the subcycle
pieces into a single piece.

Not really a true parallel epu, but can be much faster than a single
serial run in some cases.

TODO:

* clean up stdout and stderr output -- everything is just
  jumbled and repeated currently.

* Investigate doing the writes in parallel to a single file avoiding
  the intermediate writes and the subsequent join.

* Better error detection.  subcycling within subcyling if needed...