EXODUS: Add ex_{get|put}_partial_coord_component() functions
In parallel runs, there is an ambiguity in the
ex_{get|put}_partial_coord() functions.  If any of the x_coor, y_coor,
or z_coor arrays are NULL, then that component is skipped in the
get|put call.  This is done so that the client can read single
compoents at a time instead of all `ndim` components.

However, in a parallel run, it is ambiguous as to whether the NULL
array is due to only reading a single component as in:

      ex_get_partial_coord(exoid, start, count, NULL, y_coor, NULL);

which wants to read only the `y_coor`.  Or the case in which a
procesor may not have any nodes on a specific processor and therefore
passes in a NULL array since no memory needed to store zero
coordinates.

IF collective IO is being done, then we can't have a processor
skipping the call to the lower level netcdf functions.

There is now a ex_get_partial_coord_component() function which will
read the specified coordinate unambiguously.