EXODUS: Change is_mpiio to is_hdf5
NetCDF used to support a mpiio output format several versions ago.
It was then deprecated and exodus then mapped it over to using
hdf5.  In recent NetCDF versions, they map the mpiio over to
PNetCDF (which makes sense).  To avoid confusion, we change the
is_mpiio flag to is_hdf5 which is NetCDF-4 format with parallel io.