IOSS: cgns - do not need sandia mods to cgns library
Based on discussions with CGNS developers, there is
a better method for provding both parallel and serial
access to CGNS files in the same execution and it is
no longer necessary to patch the library.

The method is as follows:
* Parallel access -- same as always
  * call `cgp_mpi_comm(util().communicator())`
  * call `cgp_open()`
* Serial access in a parallel program:
  * Call `cgp_mpi_comm(MPI_COMM_SELF)`
  * call `cgp_open()`
  * call `cgp_mpi_comm(util().communicator())`
* Serial access in serial program:
  * call `cg_open()`

This has been tested on Mac and Linux, but will be tested more
extensively on larger proc counts soon.