EXODUS: Better functioning block_param output for parallel
In a parallel N->1 case, the old ex_put_block_params() function
was encountering a race condition since it would output block ids
one at a time and would verify that block ids were unique.

In the N->1 case, a rank could output the id slightly before another
rank in which case the other rank would flags a duplicate id error.

Changed such that if the type of all blocks being output is the
same and all defined blocks are being output, it now calls the
ex_put_homogenous_block_params() function which does some of the
outputs all at once instead of singly which works better for
N->1 parallel.